pass offset to BE

+++ frontend/lib/services/solver_service.dart
@@ -17,7 +17,7 @@ class SolverService {
           'color': _mapComponent(r.relationship.component),
           'operation': _mapOperator(r.relationship.operator),
           'indexes': [r.sourceShapeIndex, r.targetShapeIndex],
-          'offset': r.relationship.offset.toInt(),
+          'offset': r.relationship.offset,
         }).toList(),
       };
